+++ CHANGES.md
@@ -1,5 +1,7 @@
 ## 0.0.49
 
+-   Set CSW connector MD_Metadata dateStamp as issue date
+
 ## 0.0.48
 
 -   Add No results label when there are no results in organisation and location facets.